Group daytype

+++ src/app/components/self/self-daytype/self-daytype.component.ts
@@ -5,13 +5,17 @@ import { ConfirmationService, MenuItem, MessageService } from 'primeng/api';
 import { AppConfig } from 'src/app/config/config';
 import { InitialCurrent } from 'src/app/config/initial_current';
 import { DaytypeModels } from 'src/app/models/attendance/daytype';
+import { AccountModel } from 'src/app/models/self/account';
 import { cls_MTReqdocumentModel } from 'src/app/models/self/cls_MTReqdocument';
 import { cls_TRTimedaytypeModel } from 'src/app/models/self/cls_TRTimedaytype';
+import { TRAccountModel } from 'src/app/models/self/traccount';
 import { ReasonsModel } from 'src/app/models/system/policy/reasons';
 import { TimecardService } from 'src/app/services/attendance/timecards.service';
+import { AccountServices } from 'src/app/services/self/account.service';
 import { TimeDaytypeServices } from 'src/app/services/self/timedaytype.service';
 import { ReasonsService } from 'src/app/services/system/policy/reasons.service';
 declare var reqdaytype: any;
+interface Status { name: string, code: number }
 @Component({
   selector: 'app-self-daytype',
   templateUrl: './self-daytype.component.html',
@@ -33,6 +37,7 @@ export class SelfDaytypeComponent implements OnInit {
     private reasonService: ReasonsService,
     private timedaytypeService: TimeDaytypeServices,
     private timecardService: TimecardService,
+    private accountServie: AccountServices,
     private router: Router,
   ) { }
   fileToUpload: File | any = null;
@@ -41,35 +46,70 @@ export class SelfDaytypeComponent implements OnInit {
   daytype_newselected: DaytypeModels = new DaytypeModels();
   items: MenuItem[] = [];
   items_attfile: MenuItem[] = [];
+  namedis: string = "worker_detail_en"
   reason_list: ReasonsModel[] = [];
   reasonselected: ReasonsModel = new ReasonsModel();
   timedaytype_list: cls_TRTimedaytypeModel[] = [];
   selectedtimedaytype: cls_TRTimedaytypeModel = new cls_TRTimedaytypeModel();
   selectedreqdoc: cls_MTReqdocumentModel = new cls_MTReqdocumentModel();
+  account_list: TRAccountModel[] = [];
+  account_list_source: TRAccountModel[] = [];
+  account_list_dest: TRAccountModel[] = [];
+  selectedAccount: TRAccountModel = new TRAccountModel();
+  start_date: Date = new Date();
+  end_date: Date = new Date();
+  status_list: Status[] = [{ name: this.langs.get('wait')[this.selectlang], code: 0 }, { name: this.langs.get('finish')[this.selectlang], code: 3 }, { name: this.langs.get('reject')[this.selectlang], code: 4 }];
+  status_select: Status = { name: this.langs.get('wait')[this.selectlang], code: 0 }
   public initial_current: InitialCurrent = new InitialCurrent();
   doGetInitialCurrent() {
     this.initial_current = JSON.parse(localStorage.getItem(AppConfig.SESSIONInitial) || '{}');
     if (!this.initial_current.Token) {
       this.router.navigateByUrl('login');
     }
+    this.start_date = new Date(`${this.initial_current.PR_Year}-01-01`);
+    this.end_date = new Date(`${this.initial_current.PR_Year}-12-31`);
     this.selectlang = this.initial_current.Language;
     if (this.initial_current.Language == "TH") {
       this.reasonedis = "reason_name_th"
       this.daytypeedis = "daytype_name_th"
+      this.namedis = "worker_detail_th"
+    }
+    if (this.initial_current.Usertype == "GRP") {
+      this.doLoadAccount();
+    } else {
+      this.doLoadTimedaytype();
     }
   }
   ngOnInit(): void {
     this.doGetInitialCurrent();
     this.doLoadMenu();
-    this.doLoadTimedaytype();
     this.doLoadReason();
     this.doLoadPolDaytype();
+  }
+  Search() {
+    this.doLoadTimedaytype();
+  }
+  doLoadAccount() {
+    var tmp = new AccountModel();
+    tmp.account_user = this.initial_current.Username;
+    tmp.account_type = this.initial_current.Usertype;
+    this.accountServie.account_get(tmp).then(async (res) => {
+      res[0].worker_data.forEach((obj: TRAccountModel) => {
+        obj.worker_detail_en = obj.worker_code + " : " + obj.worker_detail_en;
+        obj.worker_detail_th = obj.worker_code + " : " + obj.worker_detail_th;
+      });
+      this.account_list = await res[0].worker_data;
+      this.selectedAccount = res[0].worker_data[0];
+      this.doLoadTimedaytype();
+    });
   }
   doLoadTimedaytype() {
     this.timedaytype_list = [];
     var tmp = new cls_TRTimedaytypeModel();
-    tmp.timedaytype_workdate = new Date(`${this.initial_current.PR_Year}-01-01`)
-    tmp.timedaytype_todate = new Date(`${this.initial_current.PR_Year}-12-31`)
+    tmp.timedaytype_workdate = this.start_date;
+    tmp.timedaytype_todate = this.end_date;
+    tmp.status = this.status_select.code;
+    tmp.worker_code = this.selectedAccount.worker_code;
     this.timedaytypeService.timedaytype_get(tmp).then(async (res) => {
       res.forEach((elm: any) => {
         elm.timedaytype_workdate = new Date(elm.timedaytype_workdate)
@@ -182,6 +222,8 @@ export class SelfDaytypeComponent implements OnInit {
         label: this.langs.get('new')[this.selectlang],
         icon: 'pi pi-fw pi-plus',
         command: (event) => {
+          this.account_list_source = [];
+          this.account_list_dest = [];
           this.selectedtimedaytype = new cls_TRTimedaytypeModel();
           this.doLoadDaytypeOld();
           this.reasonselected = this.reason_list[0]
@@ -190,6 +232,11 @@ export class SelfDaytypeComponent implements OnInit {
           this.selectedtimedaytype.timedaytype_new = this.daytype_list[0].daytype_code
           this.selectedtimedaytype.reason_code = this.reason_list[0].reason_code
           // this.selectedtrtimeot.location_code = this.location_list[0].location_code
+          if (this.initial_current.Usertype == "GRP") {
+            this.account_list.forEach((obj: TRAccountModel) => {
+              this.account_list_source.push(obj)
+            })
+          }
           this.showManage()
         }
 
@@ -327,14 +374,50 @@ export class SelfDaytypeComponent implements OnInit {
     this.displayManage = true
   }
   Delete() {
-    this.doDeleteTimedaytype(this.selectedtimedaytype)
+    this.confirmationService.confirm({
+      message: this.langs.get('confirm_delete_doc')[this.selectlang],
+      header: this.langs.get('title_daytype')[this.selectlang],
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.doDeleteTimedaytype(this.selectedtimedaytype)
+      },
+      reject: () => {
+      }
+    });
   }
   Save() {
-    if (this.selectedtimedaytype.timedaytype_doc === "") {
-      this.selectedtimedaytype.timedaytype_doc = "DAYTYPE_" + this.datePipe.transform(new Date(), 'yyyyMMddHHmmss');
-    }
-    // console.log(this.selectedtimedaytype)
-    this.doRecordTimedaytype([this.selectedtimedaytype])
+    this.confirmationService.confirm({
+      message: this.langs.get('confirm_doc')[this.selectlang],
+      header: this.langs.get('title_daytype')[this.selectlang],
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        if (this.initial_current.Usertype == "GRP" && !this.edit_data) {
+          let data_doc: cls_TRTimedaytypeModel[] = []
+          this.account_list_dest.forEach((obj: TRAccountModel, index) => {
+            var tmp: cls_TRTimedaytypeModel = new cls_TRTimedaytypeModel();
+            tmp.timedaytype_id = this.selectedtimedaytype.timedaytype_id;
+            tmp.timedaytype_doc = "DAYTYPE_" + (Number(this.datePipe.transform(new Date(), 'yyyyMMddHHmmss')) + index);
+            tmp.timedaytype_workdate = this.selectedtimedaytype.timedaytype_workdate;
+            tmp.timedaytype_todate = this.selectedtimedaytype.timedaytype_todate;
+            tmp.timedaytype_new = this.selectedtimedaytype.timedaytype_new;
+            tmp.reason_code = this.selectedtimedaytype.reason_code;
+            tmp.timedaytype_note = this.selectedtimedaytype.timedaytype_note;
+            tmp.reqdoc_data = this.selectedtimedaytype.reqdoc_data;
+            tmp.worker_code = obj.worker_code;
+            data_doc.push(tmp)
+          })
+          this.doRecordTimedaytype(data_doc)
+        } else {
+          if (this.selectedtimedaytype.timedaytype_doc === "") {
+            this.selectedtimedaytype.timedaytype_doc = "DAYTYPE_" + this.datePipe.transform(new Date(), 'yyyyMMddHHmmss');
+          }
+          // console.log(this.selectedtimedaytype)
+          this.doRecordTimedaytype([this.selectedtimedaytype])
+        }
+      },
+      reject: () => {
+      }
+    });
   }
   getFullStatus(code: string) {
     let status = ""
